src/backend/requirements.txt
accelerate==0.33.0
acvl_utils==0.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
batchgenerators==0.25
batchgeneratorsv2==0.2.1
beartype==0.18.5
bitsandbytes==0.42.0
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
CoLT5-attention==0.11.1
connected-components-3d==3.18.0
contourpy==1.2.1
cycler==0.12.1
dataclasses==0.6
datasets==2.21.0
dicom2nifti==2.4.11
dill==0.3.8
distro==1.9.0
dynamic_network_architectures==0.3.1
einops==0.8.0
einops-exts==0.0.4
einx==0.3.0
exceptiongroup==1.2.2
fastapi==0.112.1
fft-conv-pytorch==1.2.0
filelock==3.15.4
flake8==7.1.1
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gensim==4.3.3
googletrans==3.0.0
graphviz==0.20.3
grpclib==0.4.7
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
hyperframe==5.2.0
idna==2.10
imagecodecs==2024.6.1
imageio==2.35.0
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
libcst==1.1.0
linecache2==1.0.0
local-attention==1.9.15
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mccabe==0.7.0
mdurl==0.1.2
MedCLIP==0.0.3
medpalm==0.2.0
modal==0.64.52
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.3
nibabel==5.2.1
nltk==3.9.1
nnunetv2==2.5.1
numpy==1.26.4
openai==0.28.0
openapi==2.0.0
opencv-python==4.10.0.84
p_tqdm==1.4.2
packaging==24.1
pandas==2.2.2
pathos==0.3.2
pillow==10.4.0
pluggy==1.5.0
pox==0.3.4
ppft==1.7.6.8
protobuf==4.25.4
psutil==6.0.0
pyarrow==17.0.0
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pydicom==2.4.4
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.1.2
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gdcm==3.0.24.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.20.0
rt-utils==1.2.7
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
shellingham==1.5.4
sigtools==4.0.1
SimpleITK==2.3.1
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
starlette==0.38.2
sympy==1.13.2
synchronicity==0.7.5
textaugment==2.0.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tifffile==2024.8.10
timm==1.0.8
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==2.4.0
TorchFix==0.5.0
torchvision==0.19.0
TotalSegmentator==2.4.0
tqdm==4.66.5
traceback2==1.4.0
transformers==4.31.0
typer==0.12.4
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unittest2==1.1.0
urllib3==2.2.2
uvicorn==0.30.6
vector-quantize-pytorch==1.15.6
watchfiles==0.23.0
wget==3.2
wrapt==1.16.0
xvfbwrapper==0.2.9
xxhash==3.5.0
yacs==0.1.8
yarl==1.9.4
zetascale==2.7.0
